refactor live mode
